Revert earlier package refactor
